tests: matrix: factor out MainTest
